requirements.txt
streamlit==1.31.0
requests==2.31.0
beautifulsoup4==4.12.3
langchain==0.1.6
faiss-cpu==1.7.4
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.2.0
numpy==1.26.3